Clean up view helpers, aka fix bugs by following convention
* don't nest view helpers in module for no reason
* rename view helpers per the convention
* add explicit csrf tags helper instead of applicationhelper
* isolate_namespace !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
