--- app/helpers/adminsimple/modules/application_helper.rb
@@ -1,8 +1,0 @@
-module Adminsimple::Modules::ApplicationHelper
-  def rescued_csrf_meta_tags
-    csrf_meta_tags
-  rescue ArgumentError
-    request.reset_session
-    csrf_meta_tags
-  end
-end
