+++ app/helpers/adminsimple/rescued_csrf_meta_tags_helper.rb
@@ -1,0 +1,10 @@
+module Adminsimple
+  module RescuedCsrfMetaTagsHelper
+    def rescued_csrf_meta_tags
+      csrf_meta_tags
+    rescue ArgumentError
+      request.reset_session
+      csrf_meta_tags
+    end
+  end
+end
